+++ src/bracket_operators.jl
@@ -1,17 +1,17 @@
-@inline function inc(j, n)
-    j + 1 > n ? j - 1 : j + 1
-end
+"""
+    _apply_P_h!(Pf, f, h, ci, li, h₁, h₂)
 
-@inline function dec(j, n)
-    j - 1 < 1 ? j + 1 : j - 1
-end
-
-### 2D Poisson Bracket in operator form: f ↦ [f,h]
-# ci and li are cartesian / linear indices for f. h₁ and h₂ are grid size in x and v
+Write Arakawa's Jacobian `[f, h]` into `Pf`, matrix-free, on a doubly periodic grid with
+`CartesianIndices` `ci`, `LinearIndices` `li` and spacings `h₁` and `h₂`. The vectors `Pf`,
+`f` and `h` hold one value per node, in the order of `li`.
+"""
 function _apply_P_h!(
         Pf::AbstractVector, f::AbstractVector, h::AbstractVector, ci, li, h₁, h₂)
     nx, ny = size(ci)
-    length(Pf) == length(f) == length(h) || throw(DimensionMismatch())
+    Base.require_one_based_indexing(Pf, f, h)
+    length(Pf) == length(f) == length(h) == nx * ny || throw(DimensionMismatch(
+        "_apply_P_h! on a $nx × $ny grid needs vectors of length $(nx * ny), got " *
+        "$(length(Pf)), $(length(f)) and $(length(h))"))
     @inbounds for ij in li
         i, j = Tuple(ci[ij])
         i₋ = mod1(i-1, nx)
@@ -42,8 +42,13 @@ function _apply_P_h!(
     end
 end
 
-### 2D Poisson Bracket in operator form: f ↦ [f,v²/2 + ϕ]
-# ci and li are cartesian / linear indices for f. h₁ and h₂ are grid size in x and v. ϕ is defined on x.
+"""
+    _apply_P_ϕ!(Pf, f, v, ϕ, ci, li, h₁, h₂)
+
+Write `[f, v²/2 + ϕ]` into `Pf`, as `_apply_P_h!` does for the Hamiltonian
+`h[i, j] = ϕ[i] + v[j]^2 / 2`, from the velocity nodes `v` of length `nv` and the potential `ϕ`
+of length `nx`, without forming `h`.
+"""
 function _apply_P_ϕ!(Pf::AbstractVector, f::AbstractVector,
         v::AbstractVector, ϕ::AbstractVector, ci, li, h₁, h₂)
     nx, nv = size(ci)
@@ -77,75 +82,5 @@ function _apply_P_ϕ!(Pf::AbstractVector, f::AbstractVector,
                f[li[i₊, j₋]] * (ϕ[i₊] + 0.5 * v[j]^2 - ϕ[i] - 0.5 * v[j₋]^2)
         )
         Pf[ij] = 1/(12 * h₁ * h₂) * (jpp + jpc + jcp)
-    end
-end
-
-### Collision Operator
-function _apply_C!(Cf::AbstractVector, f::AbstractVector, v::AbstractVector,
-        u::AbstractVector, ε::AbstractVector, ci, li, h₁, h₂)
-    nx, nv = size(ci)
-    length(Cf) == length(f) == nx*nv || throw(DimensionMismatch())
-    length(v) == nv || throw(DimensionMismatch())
-    length(u) == length(ε) == nx || throw(DimensionMismatch())
-    @inbounds for ij in li
-        i, j = Tuple(ci[ij])
-        i₋ = mod1(i-1, nx)
-        j₋ = mod1(j-1, nv)
-        i₊ = mod1(i+1, nx)
-        j₊ = mod1(j+1, nv)
-
-        Cf[ij] = (ε[i] - u[i]^2) * (f[li[i, j₋]] - 2f[ij] + f[li[i, j₊]]) / h₂^2
-        Cf[ij] += ((v[j₊] - u[i]) * f[li[i, j₊]] - (v[j₋] - u[i]) * f[li[i, j₋]]) / (2h₂)
-    end
-end
-
-function _apply_Cρ!(Cf::AbstractVector, f::AbstractVector, v::AbstractVector,
-        ρu::AbstractVector, ρε::AbstractVector, ρ::AbstractVector, ci, li, h₁, h₂)
-    nx, nv = size(ci)
-    length(Cf) == length(f) == nx*nv || throw(DimensionMismatch())
-    length(v) == nv || throw(DimensionMismatch())
-    length(ρu) == length(ρε) == length(ρ) == nx || throw(DimensionMismatch())
-    @inbounds for ij in li
-        i, j = Tuple(ci[ij])
-        i₋ = mod1(i-1, nx)
-        j₋ = mod1(j-1, nv)
-        i₊ = mod1(i+1, nx)
-        j₊ = mod1(j+1, nv)
-
-        Cf[ij] = (ρε[i] - ρu[i]^2 / ρ[i]) * (f[li[i, j₋]] - 2f[ij] + f[li[i, j₊]]) / h₂^2
-        Cf[ij] += ((ρ[i] * v[j₊] - ρu[i]) * f[li[i, j₊]] -
-                   (ρ[i] * v[j₋] - ρu[i]) * f[li[i, j₋]]) / (2h₂)
-    end
-end
-
-function _apply_Cρ²!(Cf::AbstractVector, f::AbstractVector, v::AbstractVector,
-        ρu::AbstractVector, ρε::AbstractVector, ρ::AbstractVector, ci, li, h₁, h₂)
-    nx, nv = size(ci)
-    length(Cf) == length(f) == nx*nv || throw(DimensionMismatch())
-    length(v) == nv || throw(DimensionMismatch())
-    length(ρu) == length(ρε) == length(ρ) == nx || throw(DimensionMismatch())
-    @inbounds for ij in li
-        i, j = Tuple(ci[ij])
-        i₋ = mod1(i-1, nx)
-        j₋ = mod1(j-1, nv)
-        i₊ = mod1(i+1, nx)
-        j₊ = mod1(j+1, nv)
-
-        Cf[ij] = (ρ[i] * ρε[i] - ρu[i]^2) * (f[li[i, j₋]] - 2f[ij] + f[li[i, j₊]]) / h₂^2
-        Cf[ij] += ρ[i] * ((ρ[i] * v[j₊] - ρu[i]) * f[li[i, j₊]] -
-                   (ρ[i] * v[j₋] - ρu[i]) * f[li[i, j₋]]) / (2h₂)
-    end
-end
-
-function _apply_Δᵥ!(Cf::AbstractVector, f::AbstractVector, ci, li, h₁, h₂)
-    nx, nv = size(ci)
-    length(Cf) == length(f) == nx*nv || throw(DimensionMismatch())
-    @inbounds for ij in li
-        i, j = Tuple(ci[ij])
-        i₋ = mod1(i-1, nx)
-        j₋ = mod1(j-1, nv)
-        i₊ = mod1(i+1, nx)
-        j₊ = mod1(j+1, nv)
-        Cf[ij] = (f[li[i, j₋]] - 2f[ij] + f[li[i, j₊]]) / h₂^2
     end
 end